Add a code to download data
download_data.py
-----
This is a script to download all the data files in a directory from the webpage.
You can change the 'year' by the list. (I picked 2019 for this script)
The default is to create a file list and download all data of the year into the directory './data/year' (e.g. './data/2019')